refactor(today_feed): extract TodayFeedErrorCard and RetryCta from error_state_widget (#P4-E)
